feat: update file names and path
